+++ anchorpoint/__init__.py
@@ -3,4 +3,4 @@
 from anchorpoint.textselectors import TextQuoteSelector, TextPositionSelector
 from anchorpoint.textselectors import TextPositionSet, TextPositionSetFactory
 
-__version__ = "0.6.1"
+__version__ = "0.7.0"